+++ packages/gatsby-tinacms-json/src/use-json-form.ts
@@ -121,19 +121,25 @@ export function useJsonForm(
   return [jsonNode, form as Form]
 }
 
+/**
+ * @deprecated See https://github.com/tinacms/rfcs/blob/master/0006-form-hook-conventions.md
+ */
 export function useLocalJsonForm(
   jsonNode: JsonNode | null,
   formOptions: Partial<FormOptions<any>> = {}
-) {
+): [any, Form | null] {
   const [values, form] = useJsonForm(jsonNode, formOptions)
   usePlugins(form as any)
   return [values, form]
 }
 
+/**
+ * @deprecated See https://github.com/tinacms/rfcs/blob/master/0006-form-hook-conventions.md
+ */
 export function useGlobalJsonForm(
   jsonNode: JsonNode | null,
   formOptions: Partial<FormOptions<any>> = {}
-) {
+): [any, Form | null] {
   const [values, form] = useJsonForm(jsonNode, formOptions)
   usePlugins(
     React.useMemo(() => {