+++ packages/demo-cra/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.27-alpha.2](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.27-alpha.1...demo-cra@1.0.27-alpha.2) (2020-04-14)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.27-alpha.1](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.27-alpha.0...demo-cra@1.0.27-alpha.1) (2020-04-07)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
 ## [1.0.27-alpha.0](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.26...demo-cra@1.0.27-alpha.0) (2020-04-06)
 
 **Note:** Version bump only for package demo-cra