+++ packages/gatsby-tinacms-mdx/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.10-alpha.2](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.10-alpha.1...gatsby-tinacms-mdx@0.1.10-alpha.2) (2020-04-14)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
+## [0.1.10-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.10-alpha.0...gatsby-tinacms-mdx@0.1.10-alpha.1) (2020-04-07)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
 ## [0.1.10-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.9...gatsby-tinacms-mdx@0.1.10-alpha.0) (2020-04-06)
 
 **Note:** Version bump only for package gatsby-tinacms-mdx